acg_loc:remove inline function, correct robot pose, landmark, weird odom
Tested robot pose and landmark expression and optimization. Seems to
work alright, however, more corner that previously found with ndt_featur
A real pb is weird number for the odometry. While this seems not to
influence the optimization much, I'm not to sure why the odom is what it
is. I must try the actual number and see if that change anything.
Ask Daniel about it also.

Best regards