acg_loc: refactorisation of function in ACG + localization working in EM
Awesome, it works as it is without having to modify the localization
code. Really surprising. I also refactored some code from ACG to
ACGLocalization -> TODO transfer the change back to ACG.
I'm adding localization poses in the graph. Corner observation from
the localization pose is still to be tested.

Malcolm