acg_loc: loc covariance in link edges, corner cov and refactor
Now the link edge use the MCL covariance as a covariance. That should be
more logical considering that they are observed in the prior using MCL.
The corner covariance can be use BUT it crashes the optimization. My
guess is that it is due to the fact that they cov matrix are close to 0.

Malcolm