acg_loc: update prior in ndt_mcl + use covariance to find match (bad)
Update the mcl during the optimization! Works top !
However, I'm trying to find a clever way to use the mcl covariance to
find observation to prior and the pb is that the x and y is "too"
certain compared to reality... I need to take the angle covariance
inconsideration probably but it's not working for now.

Malcolm